Update operations.py
Removed the error handling on GetLog for NOT_FOUND, as I don't think you expect this to normally occur the same way it would for a regular get (i.e. it may be intuitive for get to return None if no associated key was found, but it isn't quite as intuitive for the getLog command to return None if the drive faults on getting a log and returns not found).